requirements.txt
moviepy==1.0.3
openai-whisper==20231117
requests==2.31.0
numpy==1.23.5